### 2024-05-13 #6750 @editorialbot
[PRE REVIEW]: PyCVI: A Python package for internal Cluster Validity Indices, compatible with time-series data
**Submitting author:** @nglm (<a href="http://orcid.org/0000-0002-8473-1673">Natacha Galmiche</a>)
**Repository:** https://github.com/nglm/pycvi
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
