### 2024-05-21 #6750 @nglm
Hi @crvernon , a combination of public holiday and sickness delayed this response. Thank you for your patience!

- The "additional edge cases" are all in the file [cvi_func.py](https://github.com/nglm/pycvi/blob/master/pycvi/cvi_func.py). For some CVIs implemented in this file, there are some `if`-statements related to specific numbers of clusters $k$ (typically $k=1$) that are our own addition and were not treated in the original CVIs and which are outside our main goal for this current submission to JOSS. They correspond to the lines 244-245, 359-372, 480-521, 563-564,894-895.
- Our progress on the other paper is not significant enough so that we can aim for a specific venue and deadline, but we are typically aiming for one of these **non-software** peer-reviewed venues:
  - [European Conference on Artificial Intelligence (ECAI)](https://www.ecai2024.eu/)
  - [Conference on Knowledge Discovery and Data Mining (KDD)](https://kdd.org/conferences)
  - [International Conference on Machine Learning (ICML)](https://icml.cc/)
  - [International Joint Conference on Artificial Intelligence (IJCAI)](https://www.ijcai.org/)
  - [Artificial Intelligence and Statistics Conference (AIStats)](https://aistats.org/)
  - [The Association for the Advancement of Artificial Intelligence (AAAI)](https://aaai.org/)
  - [The Conference on Uncertainty in Artificial Intelligence (UAI)](https://www.auai.org/)
  - [Northern Lights Deep Learning Conference  NLDL](https://www.nldl.org/)
