### 2024-05-22 #6750 @nglm
@lrnv  based on previous JOSS submissions on clustering, I could recommend:
- martinfleis for his work on [martinfleis/clustergram](https://github.com/martinfleis/clustergram)
- ajgates42 and yy for their work on [Hoosier-Clusters/clusim](https://github.com/Hoosier-Clusters/clusim)
