### 2024-06-03 #6750 @lrnv
@TahiriNadia, @peekxc, @robcaulk, @wob86, @branii, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
