### 2024-06-04 #6750 @lrnv
@laurent.vouriot , would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: joss.readthedocs.io/en/latest/review_criteria.html
