### 2024-06-17 #6841 @lrnv
@robcaulk, @wob86, @laurent-vouriot, how is it going since last week ? Do not hesitate to tell me if there is any questions that i could help with.
