### 2024-06-18 50200b8
added a script illustrating importance DBA
