### 2024-06-29 #6841 @robcaulk
> @robcaulk, @wob86, @laurent-vouriot, how is it going since last week ? Do not hesitate to tell me if there is anything i might be able to help with.

Hello, I just got around to working on this today. There are some issues preventing moving forward for now:
https://github.com/nglm/pycvi/issues/4
