### 2024-06-29 #6841 @robcaulk
Some other questions I have regarding the paper:

*Please note, that I would not consider myself a scholar in clustering. I am, however, a "power-user" of scikit-learn pipelines/clustering, so I am bringing a user perspective to this review. I would appreciate if you could correct any of my misunderstandings around the use of variational information and CVIs, as my questions may be naive. My aim here is to evaluate this repository from a practical perspective as well as from a tool-use perspective. *

## Primary question

My principal question here is related to the "Substantial scholarly effort" checkbox. Which relates to likelihood of citation in academia.

If I understand correctly, this library is an aggregation of many existing CVI methods. I'm not convinced that an academic would need to cite this library instead of simply citing the individual clustering/CVI methods directly. Perhaps the argument is that an academic would cite both the method and the library. Ok, but I think the library in that case needs to bring something novel, like some "auto-selection" of the CVI based on some analytics of the clusters, or something (please see my note above about how I am not a scholar on clustering methods). It seems they are doing something with the variation of information (VI), but if I understood correctly, it is not used for automatically choosing the best CVI.

It introduces my next question: "why did they elect against contributing more CVI methods to scikit-learn?" In other words "what exactly forced the need of a new library?" If I understand correctly, scikit-learn already has 3 CVI methods, while this current library has 12. Why not simply augment the offering at scikit-learn, which opens up these algorithms to a much wider user base?

In general, I think they need to highlight exactly what novelty this repository brings. I believe the novelty/answer to my above questions is hidden somewhere in this statement from the paper:

> Unfortunately, DTW and DBA are not compatible with the libraries mentioned above.

and

> Then, in PyCVI their 46 definition is extended in order to make them compatible with DTW and DBA in addition to
47 non time-series data.

But I cannot find any support or rationale for *why* DTW and DBA are not compatible with scikit-learn or *why* this was a difficult implementation that needed a separate library. Again, I am not a scholar in clustering, so maybe this is obvious to the authors, but i think a couple sentences on this contribution is key to the "Scholarly Contribution" of this library.

## Secondary question

Another question that is not clear to me is the claim that:

> PyCVI is entirely compatible with scikit-learn clustering pipelines

I was unable to locate an example showing the use of `pycvi` in a scikit-learn `Pipeline`. So I am curious what the authors meant by this, or if they can show an example to support the claim.

## Tertiary issues

From a practical perspective, I was unable to run the `basic_usage.py`, please review the steps that I took in the issue at the target repository:

https://github.com/nglm/pycvi/issues/4
