### 2024-07-02 27721bb
testing scripts without relative imports
