### 2024-07-02 1a70a8e
clearer instructions to run scripts
