+++ README.md
@@ -30,7 +30,7 @@ Compatible with [scikit-learn](https://scikit-learn.org/stable/index.html), [sci
 
 ## Install
 
-- With poetry:
+### With poetry:
 
 ```bash
 # From PyPI
@@ -39,7 +39,7 @@ poetry add pycvi-lib
 poetry add git+https://github.com/nglm/pycvi.git
 ```
 
-- With pip:
+### With pip:
 
 ```bash
 # From PyPI
@@ -48,7 +48,7 @@ pip install pycvi-lib
 pip install git+https://github.com/nglm/pycvi.git
 ```
 
-- With anaconda:
+### With anaconda:
 
 ```bash
 # activate your environment (replace myEnv with your environment name)
@@ -58,6 +58,21 @@ conda install pip
 # install pycvi on your anaconda environment with pip
 pip install pycvi-lib
 ```
+
+### Extra dependencies
+
+In order to run the example scripts, extra dependencies are necessary. The install command is then:
+
+```bash
+# For poetry
+poetry add pycvi-lib[examples]
+# For pip and anaconda
+pip install pycvi-lib[examples]
+```
+
+Alternatively, you can manually install in your environment the packages that are necessary to run the example scripts (`matplotlib` and/or `scikit-learn-extra` depending on the example).
+
+If you wish to run the example scripts on your own computer, please follow the instructions detailed in the documentation first: [Running example scripts on your computer](https://pycvi.readthedocs.io/en/latest/examples/examples_instructions.html).
 
 ## Contribute
 
