### 2024-07-17 c080c54
Not allowing a k selection when all scores are equal
