+++ pycvi/cvi.py
@@ -227,8 +227,9 @@ class CVI():
         Does not take into account rules that are specific to a CVI,
         such as the Gap statistic, Hartigan, etc.
 
-        Returns None if no clustering could be selected (probably
-        because the CVI values were None or NaN).
+        Returns `None` if no clustering could be selected (probably
+        because the CVI values were all None or NaN or that all valid
+        and relevant scores were all equal).
 
         Parameters
         ----------
@@ -254,51 +255,80 @@ class CVI():
             cvi_type = self.cvi_type
         # For monotonous CVIs
         if cvi_type == "monotonous":
+            # Remove special cases if the score couldn't be computed
+            # either because this CVI doesn't allow this k or because
+            # the clustering model didn't converge
             scores_valid = {k: s for k,s in scores.items() if s is not None}
             selected_k = None
             max_diff = 0
 
-            list_k = list(scores_valid.keys())
-            last_relevant_score = scores_valid[list_k[0]]
-            last_relevant_k = list_k[0]
+            # Sort k in increasing order
+            list_k = sorted(scores_valid.keys())
 
-            # If it is improving compare last to current
+            # If score improving with k, then compare last to current
             # Otherwise compare current to next
             if self.improve:
+                # Take the 1st score / k as the initial relevant score / k
+                last_relevant_score = scores_valid[list_k[0]]
+                last_relevant_k = list_k[0]
                 list_k_compare = list_k[1:]
             else:
+                # Take the last score / k as the initial relevant score / k
+                last_relevant_score = scores_valid[list_k[-1]]
+                last_relevant_k = list_k[-1]
                 list_k_compare = list_k[:-1]
+
             # Find the biggest increase / drop
-            for i, k in enumerate(list_k_compare):
-                # Special case if the score couldn't be computed
-                # either because this CVI doesn't allow this k
-                # or because the clustering model didn't converge
-                if scores_valid[k] is None:
-                    continue
+            for k in list_k_compare:
+
                 # Special case for example for Hartigan where
                 # we don't use k=0 as a reference score if k=0 is more
                 # relevant than k=1
-                elif (i==0 and self.ignore0 and not self.is_relevant(
+                if (last_relevant_k==0 and self.ignore0 and not self.is_relevant(
                         scores_valid[k], k,
                         last_relevant_score, last_relevant_k)
                     ):
-                    selected_k = 1
-                    last_relevant_k = 1
+                    # Use k for next iterations
+                    last_relevant_k = k
                     last_relevant_score = scores_valid[k]
+                    # If all scores are worse than the case k=0,
+                    # choose the current k as the selected k
+                    # This value will be overwritten if there is a better k
+                    selected_k = k
+                # Otherwise,
+                # If the current score is relevant
+                # Then, update relevant k and score
                 elif (
-                    (i==0 and not self.ignore0)
-                    or self.is_relevant(
+                    self.is_relevant(
                         scores_valid[k], k,
                         last_relevant_score, last_relevant_k
                     )
                 ):
+
+                    # The final selected k is the k that maximises the diff
                     diff = abs(scores_valid[k] - last_relevant_score)
                     if max_diff < diff:
                         selected_k = k
                         max_diff = diff
+                    # If e.g. scores_valid[k] = last_relevant_score = np.inf
+                    # Then diff is NaN. If so, stop already, the selected k
+                    # is the k that gave the first np.inf diff.
+                    elif np.isnan(diff):
+                        break
 
                     last_relevant_k = k
                     last_relevant_score = scores_valid[k]
+
+                # If the score was not relevant, just ignore this k, and
[…]
