### 2024-07-18 #6841 @nglm
Thank you all again for your feedback, I am now back from holidays and started working again on this review.

In this post, I will answer to @robcaulk questions.

## Question: Why do we need PyCVI?

### No implementation of individual CVI in python

> If I understand correctly, this library is an aggregation of many existing CVI methods. I'm not convinced that an academic would need to cite this library instead of simply citing the individual clustering/CVI methods directly.

The CVIs implemented in this library are indeed all already existing, but there is no available implementation of them, which slows down further research on the topic and diminishes the quality of scripts of the general user which rely on the very few methods that are available in sklearn. You then brought up the next very natural follow-up question: "why did they elect against contributing more CVI methods to scikit-learn?", I will detail the answer to this question below.

### Auto-selection

> I think the library in that case needs to bring something novel, like some "auto-selection" of the CVI based on some analytics of the clusters, or something.

There is to this day not enough research on this topic to be able to select the right CVI for the right dataset, although some papers initiated this effort. See for example [^1] or [^2]. Another obstacle to this approach is that for this to work, one would have to actually get to have a proper "analytics" of the clusters which is not really feasible outside generated toy examples to have specific characteristics or very well studied datasets.

> It seems they are doing something with the variation of information (VI), but if I understood correctly, it is not used for automatically choosing the best CVI.

VI can only be used when one knows the true clusters, therefore it is meant here to be used on known datasets in order to study the performance of CVIs or clustering methods (in a similar approach as the one detailed in [^2]), but it can not be used in a real-life clustering setting in which we precisely don't know the ground truth.

### Sklearn not compatible with time series data, nor DTW, nor DBA

#### Shapes of the datasets in sklearn

The first naive observation is that sklearn only allows datasets with the shape `(n_samples, n_features)`. This of course could still work with time series if we either flatten the time-series first, or contribute to sklearn to allow for datasets of different shapes. The second option could be more ambitious than it seems given the size of the sklearn library and given how machine learning libraries heavily rely on having a specific shape to compute distances.  The first option, flattening the data before using sklearn as is, would create additional ugly lines which in itself could be an acceptable sacrifice but is not the main problem. Although, being a very naive issue, this is already a reason to use aeon or tslearn or sktime instead of sklearn in case of time series data.

#### Use of custom metric functions in sklearn clustering methods

First of all, very few clustering methods implemented in sklearn are actually compatible with using a custom distance function (`"metric"` parameter). For example, [KMeans](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.KMeans.html#sklearn.cluster.KMeans), [MeanShift](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.MeanShift.html#sklearn.cluster.MeanShift), [SpectralClustering](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.SpectralClustering.html#sklearn.cluster.SpectralClustering) are not while [AgglomerativeClustering](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.AgglomerativeClustering.html#sklearn.cluster.AgglomerativeClustering) and [DBSCAN](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.DBSCAN.html#sklearn.cluster.DBSCAN) are. This is by the way one reason why users might not use sklearn but scikit-learn-extra (e.g. [KMedoids](https://scikit-learn-extra.readthedocs.io/en/stable/generated/sklearn_extra.cluster.KMedoids.html#sklearn_extra.cluster.KMedoids) or [CLARA](https://scikit-learn-extra.readthedocs.io/en/stable/generated/sklearn_extra.cluster.CLARA.html#sklearn_extra.cluster.CLARA)).

In addition, this custom distance function, when available, has to take as inputs `(n_samples, n_features)` datasets. Again, while adding a few ugly lines of code, we could in theory define a distance function that first unflattens a time-series dataset (provided that we know the original shape) and before calling the actual DTW distance function.

Finally, these tricks might let us use time-series data and DTW for clustering methods that allow it, but this would still not solve the issue of the definition of the average in the case of time series data. DTW can compute pairwise distances, but can not directly compute the average of a group of time-series. If the usual definition of the average is used (i.e. using the euclidean distance), then the resulting average time-serie might be very far from representing what we would expect from the average time series. For this I can recommend the original DBA paper from Petitjean et al. [^DBA], and the corresponding [GitHub repository](https://github.com/fpetitjean/DBA) as well as the [cluster center example in PyCVI](https://pycvi.readthedocs.io/en/stable/examples/cluster_center.html). The concept of average is essential in clustering as it is often used to create clusters (by comparing the distances to the different cluster centroids) but also to evaluate clusters (by computing for example the inertia or most of CVIs). This would completely change the code of for example the KMeans or Ward algorithms. And there, time-series adapted libraries such as aeon, sktime and tslearn become strictly necessary.

## Question: Sklearn Pipeline (and back to the need of PyCVI)

I don't see where your quote comes from. My guess is that you refer to this sentence:

> PyCVI is entirely compatible with scikit-learn, scikit-learn-extra, aeon and sktime, in order to be easily integrated into any clustering pipeline in python.

Which meant something totally different, and did not refer to the Pipeline class in sklearn.

As explained above, sklearn is not compatible with time-series data, nor distances adapted to time series data. If a python user want to use clustering methods that are adapted to time series data, they have to switch to another library (e.g. aeon, sktime or tslearn), which would result in different structures in their scripts compared to when they are not working with time-series data (in which case they would use sklearn).

The first motivation of PyCVI is to fill the gap in terms of CVIs, this has nothing to do (originally) with time-series data, since the gap is present even for non-time series data. But, the same reasons that made time-series specific libraries necessary, would make a time-series compatible CVI library.

PyCVI is compatible with both time series and non time series data, with the exact same syntax and functions for time series data and non time series data. In addition, PyCVI is compatible with sklearn as well as aeon and sktime, with the exact same syntax when combined with sklearn clustering methods or aeon clustering methods. Which means that the user would use PyCVI exactly in the same way regardless of the type of data they are handling and would integrate PyCVI snippets of code exactly in the same way in their already written scripts regardless of the original library used sklearn, scikit-learn extra or aeon or sktime.

## Question: runnning examples

See my reply on the corresponding issue, please let me know if there is anything more I should do on that matter.

[^1]: Y. Liu, Z. Li, H. Xiong, X. Gao, J. Wu, and S. Wu, “Understanding and enhancement of internal clustering validation measures,” IEEE Transactions on Cybernetics, vol. 43, pp. 982–994, June 2013.
[^2]:  I. Gurrutxaga, J. Muguerza, O. Arbelaitz, J. M. P ́erez, and J. I. Mart ́ın, “Towards a standard methodology to evaluate internal cluster validity indices,” Pattern Recognition Letters, vol. 32, p. 505–515, Feb. 2011.
[^DBA]: F. Petitjean, A. Ketterlin, and P. Gan carski, “A global averaging method for dynamic time warping, with applications to clustering,” Pattern Recognition, vol. 44, pp. 678–693, Mar. 2011.
