### 2024-07-18 bb3d576
More details for each fig in paper
