+++ paper/paper.md
@@ -67,13 +67,13 @@ We experimented on 3 different cases: non time-series data [@barton2015clusterin
 
 For each case, we used a different clustering method from a different library: KMeans [@lloyd1982least] and AgglomerativeClustering [@Ward1963] from [scikit-learn](https://scikit-learn.org/stable/index.html), TimeSeriesKMeans from [aeon](https://www.aeon-toolkit.org/en/latest/index.html) and KMedoids [@Kaufman1990Partitioning] from [scikit-learn-extra](https://scikit-learn-extra.readthedocs.io/en/stable/) in order to give examples of integration with other clustering libraries. Then, for each case, we ran all the CVIs implemented in PyCVI, selected the best clustering according to each CVI and plotted the selected clustering.
 
-Finally, we computed the variation of information (VI) between each selected clustering and the true clustering (second plot of all figures). A large variation of information there indicates a poor clustering quality due to the clustering method. In \autoref{fig:kmeans} and \autoref{fig:agglo}, we can see the difference of quality when assuming the correct number of clusters between the AgglomerativeClustering and the KMeans clustering method on the non time-series data. Therefore, when the quality of the clustering selected by the CVI is poor it can then either be due to the clustering method or the CVI, hence the necessity of robust evaluation pipeline for both clustering methods and CVIs.
+Finally, we computed the variation of information (VI) between each selected clustering and the true clustering (second plot of all figures). A large variation of information there indicates a poor clustering quality due to the clustering method. In \autoref{fig:kmeans} and \autoref{fig:agglo}, we can see the difference of quality when assuming the correct number of clusters between the AgglomerativeClustering and the KMeans clustering method on the non time-series data. Therefore, when the quality of the clustering selected by the CVI is poor it can then either be due to the clustering method or due to the CVI, hence the necessity of robust evaluation pipeline for both clustering methods and CVIs.
 
 In \autoref{fig:kmeans}, since the generated clusterings are poor due to the clustering method, the poor results indicated by the histogram in \autoref{fig:kmeans_hist} gives us no information about the correct number of clusters, nor about the quality of the CVIs used. However, in \autoref{fig:agglo}, the quality of the clustering is excellent, as indicated by a null VI. The poor results shown in the corresponding histogram \autoref{fig:agglo_hist} tells us that the CVI used here are not adapted to this dataset. This was expected since most CVIs rely on the cluster center to compute a good separation between clusters. The dataset here consist of concentric circles, which means that most CVIs fail to measure how well separated the clusters actually are. This illustrates the need of further research on CVIs, which is facilitated by PyCVI, notably in the case of concentric subgroups.
 
-Similarly, in the case of time-series data, we see a difference in the quality of the clustered data assuming the correct number of clusters in figures \autoref{fig:DTW} and \autoref{fig:no_DTW}, although the same clustering method, KMeans is used. This is this time due to difference between using DTW as a distance measure compared to the euclidean distance, and the difference between using DBA to compute the average of a group of time series and using the usual average.
+Similarly, in the case of time-series data, we see a difference in the quality of the clustered data assuming the correct number of clusters in figures \autoref{fig:DTW} and \autoref{fig:no_DTW}, although the same clustering method, KMeans, is used. This is this time due to difference between using DTW as a distance measure compared to using the euclidean distance, and consequently, the difference between using DBA to compute the average of a group of time series and using the usual average.
 
[…]
