+++ paper/paper.md
@@ -59,7 +59,7 @@ PyCVI then tries to fill that gap by implementing 12 state-of-the-art internal C
 
 ![KMeans clustering on time-series data, with DTW, selected number of clusters according to all implemented CVIs. \label{fig:DTW-hist}](./UCR_data_DTW_TimeSeriesKMeans-histogram.png){width="50%"}
 
-![KMedoids clustering on time-series data, without DTW, all implemented CVIs.\label{fig:no_DTW-hist}](./UCR_data_no_DTW_KMedoids.png)
+![KMedoids clustering on time-series data, without DTW, all implemented CVIs.\label{fig:no_DTW}](./UCR_data_no_DTW_KMedoids.png)
 
 ![KMedoids clustering on time-series data, without DTW, selected number of clusters according to all implemented CVIs. \label{fig:no_DTW-hist}](./UCR_data_no_DTW_KMedoids-histogram.png){width="50%"}
 
@@ -73,7 +73,7 @@ In \autoref{fig:kmeans}, since the generated clusterings are poor due to the clu
 
 Similarly, in the case of time-series data, we see a difference in the quality of the clustered data assuming the correct number of clusters in figures \autoref{fig:DTW} and \autoref{fig:no_DTW}, although the same clustering method, KMeans, is used. This is this time due to difference between using DTW as a distance measure compared to using the euclidean distance, and consequently, the difference between using DBA to compute the average of a group of time series and using the usual average.
 
-However, both time series cases suggest that there are only two clusters, as indicated by figures \autoref{fig:DTW-hist} and \autoref{fig:no_DTW-hist}. And indeed, when looking at the true labels, we see that the purple and green clusters are very similar, one being slightly more noisy than the other. In addition, the red one and the blue one are also very similar, except for the time steps between 50 and 100 (one goes up while the other goes down). This illustrates the possibility of multiple relevant interpretations of clustered data, and ideally CVIs could be used to assess the relevance of the different interpretations. In figure \autoref{fig:DTW-hist}, we see for example that 2 clusters is the most relevant interpretation, which is an acceptable interpretation given our analysis of the clusters, but is still probably less relevant than interpreting the data has consisting of 4 clusters, knowing that this is the actual number of clusters in the dataset. Thus, ideally, figures \autoref{fig:DTW} and \autoref{fig:DTW-hist} should have to high bars, one for 2 clusters and one for 4 clusters. However, most CVIs seem to fail to identify 4 clusters, yielding a very low bar for 4 in both figures \autoref{fig:DTW} and \autoref{fig:DTW-hist}. This suggests that future work on CVIs could significantly improve the quality of selected clusterings and could give us more insights on the relevant interpretation of the data in case of ambiguous clusters.
[…]
