### 2024-08-15 f23d235
clearer _check_list_of_dict
