### 2024-08-15 c7786dc
starting CVIAggregator
