### 2024-08-15 68ebac8
clearer _check_list_of_dict
