### 2024-08-16 ca1ec4b
Added CVIAggregator class
