### 2024-08-16 be921b6
adapted compute_all_scores to CVI and CVIAggregator
