+++ pycvi/tests/test_compute_scores.py
@@ -4,6 +4,8 @@ import pytest
 from aeon.clustering import TimeSeriesKMeans
 from sklearn.cluster import KMeans
 from sklearn.preprocessing import StandardScaler
+
+from ._utils import _aux_check_float
 from ..cvi import Inertia, GapStatistic
 
 from ..datasets._mini import mini
@@ -53,22 +55,23 @@ def test_compute_subscores():
 
         # DTW case
         dist = _compute_subscores("inertia", data, [c1, c2], "inertia", f_inertia)
-        assert (type(dist) == float or type(dist) == np.float64)
+        # float
+        _aux_check_float(dist, or_None=False)
         dist = _compute_subscores("max_inertia", data, [c1, c2], "inertia", f_inertia)
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_subscores("list_inertia", data, [c1, c2], "inertia", f_inertia)
         assert type(dist) == list
-        assert (type(dist[0]) == float or type(dist[0]) == np.float64)
+        _aux_check_float(dist[0], or_None=False)
 
         # Non DTW case
         data = data.reshape(N, -1)
         dist = _compute_subscores("inertia", data, [c1, c2], "inertia", f_inertia)
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_subscores("max_inertia", data, [c1, c2], "inertia", f_inertia)
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_subscores("list_inertia", data, [c1, c2], "inertia", f_inertia)
         assert type(dist) == list
-        assert (type(dist[0]) == float or type(dist[0]) == np.float64)
+        _aux_check_float(dist[0], or_None=False)
 
 def test_compute_score():
     """
@@ -82,30 +85,30 @@ def test_compute_score():
 
         # DTW case
         dist = _compute_score("inertia", data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_score("max_inertia", data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_score("list_inertia", data, [c1, c2])
         assert type(dist) == list
-        assert (type(dist[0]) == float or type(dist[0]) == np.float64)
+        _aux_check_float(dist[0], or_None=False)
         dist = _compute_score(Inertia(), data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_score(silhouette, data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
 
         # Non DTW case
         data = data.reshape(N, -1)
         dist = _compute_score("inertia", data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_score("max_inertia", data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
+        _aux_check_float(dist, or_None=False)
         dist = _compute_score("list_inertia", data, [c1, c2])
         assert type(dist) == list
-        assert (type(dist[0]) == float or type(dist[0]) == np.float64)
+        _aux_check_float(dist[0], or_None=False)
         dist = _compute_score(Inertia(), data, [c1, c2])
-        assert (type(dist) == float or type(dist) == np.float64)
[…]
