+++ pycvi/tests/_utils.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+def _aux_check_float(value, or_None=False):
+    if or_None:
+        assert (type(value) in [float, np.float64, type(None)])
+    else:
+        assert (type(value) in [float, np.float64])
