### 2024-08-16 f651fa6
added test_cviaggregator_init
