### 2024-08-16 5e021d0
fix output types diameter, inertia, etc
