+++ pycvi/compute_scores.py
@@ -90,7 +90,7 @@ def _f_generalized_var(
     :rtype: float
     """
     if len(cluster) == 1:
-        return 0
+        return 0.
     else:
         # a dxd array, representing the sample covariance
         sample_cov = np.cov(cluster, rowvar=False)
@@ -114,7 +114,7 @@ def _f_med_dev_centroid(
     :rtype: float
     """
     if len(cluster) == 1:
-        return 0
+        return 0.
     else:
         centroid = np.expand_dims(compute_center(cluster), 0)
         dist = f_cdist(cluster, centroid, **dist_kwargs )
@@ -135,7 +135,7 @@ def _f_mean_dev_med(
     :rtype: float
     """
     if len(cluster) == 1:
-        return 0
+        return 0.
     else:
         centroid = np.expand_dims(compute_center(cluster), 0)
         dist = f_cdist(cluster, centroid, **dist_kwargs )
@@ -158,7 +158,7 @@ def _f_med_dev_med(
     :rtype: float
     """
     if len(cluster) == 1:
-        return 0
+        return 0.
     else:
         return np.median(cdist(
             cluster,
@@ -188,7 +188,7 @@ def f_diameter(
         The diameter of the cluster.
     """
     if len(cluster) == 1:
-        return 0
+        return 0.
     else:
         pdist = f_pdist(cluster, dist_kwargs)
         return np.amax(pdist)
