### 2024-08-16 e6cb41f
fix reduction kwarg in Diameter and Inertia CVI
