### 2024-08-16 7e636bc
fix cviaggregator.select method
