### 2024-08-16 84ecd90
added test_cviaggregator - all tests passed
