### 2024-08-19 5652b3e
adding CVIAggregator example
