### 2024-08-19 1a3a8e9
updated paper (CVIAggregator, emphasize contributions)
