### 2024-08-19 a37107b
updated paper (explain VI)
