### 2024-08-19 #6841 @nglm
Once again thank you so much for the very useful feedback! I have now updated the library, the paper and the documentation in an attempt to answer @laurent-vouriot questions and concerns. This post tries to summarize the different updates.

## On updating the paper

- Clearer explanation of the Variation of information
- One more example, with successful CVIs: added one example featuring `pycvi.cvi.CVIAggregator`. The example was added to the documentation and to the paper, showing successful CVIs for a given dataset (see details on `pycvi.cvi.CVIAggregator` below).
- Stronger focus on the importance of time series compatibility.

## On adding a scorer-like wrapper

By scorer-like wrapper, I mean (based on what I understand from your post) a sklearn-like scrorer as described [here](https://scikit-learn.org/stable/modules/model_evaluation.html#custom-scorer-objects) and [there](https://scikit-learn.org/stable/modules/model_evaluation.html#how-to-build-a-scorer-from-scratch) and which can be summarized as a callable that:

- can be called with parameters `(estimator, X, y)`,
- returns a floating point number that quantifies the estimator prediction quality on X, with reference to y, using the convention that higher numbers are better.

### Objectives:

1. Make it easier to integrate PyCVI's CVIs into sklearn pipelines, notably those using [GridSearchCV](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.GridSearchCV.html), [RandomizedSearchCV](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.RandomizedSearchCV.html) and [cross_validate](https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_validate.html).
2. Make it easier to use mulitple CVIs at a time.
3. Make PyCVI pipelines more general and re-usable by creating an object that should always be maximized from each implemented CVI.

### Challenges / Problems:

1. No particular challenge regarding this objective.
2. There is no technical challenge about this objective, but rather a "design choice" against it. We believe that some datasets are ambiguous and that multiple relevant interpretations of each of them (e.g. multiple relevant interpretations on the numbers of clusters) can coexist and should not be hidden to the user by showing only the "best" one and hidding the other potentially relevant ones (this is typically the case with datasets with subclusters, noisy clusters, clusters with different variance or number of datapoints, etc.). This point of view is rather specific to clustering tasks because we don't have any ground truth to tell us what is *the* correct interpretation, although there exists also similar issues with fuzzy classification (classification problem in which each input can belong to several labels "to some degree") and multi-label classification (classification problem in which each input can have several true labels). Those edge cases of classification (and our viewpoint on clustering) are not taken into account in sklearn and/or out of the scope of sklearn. Our recommendation to face this issue is to plot the number of CVIs that voted for a given number of cluster with respect to a given range of number of clusters, as we did in our example ["Selecting the number of clusters k"](https://pycvi.readthedocs.io/en/latest/examples/select_k.html)
3. Many CVIs are relative and/or using a threshold and/or using a specific selection rule given a sequence of cvi value for different clusterings. In such cases, it is impossible to output a float to be maximised for some given clusterings. The cvi value cannot be further exploited other than to select *the* best clustering among a set of clusterings. This particularity/difficulty is however mitigated by our "select" method, common to all implemented CVIs, that selects the best clustering given computed CVI values using the same user syntax regardless of the CVI used (see the line further down in the ["full example"](https://pycvi.readthedocs.io/en/latest/examples/full_example.html) "```k_selected = cvi.select(scores)```").  Note that we have precisely attempted to extend some existing CVIs to make their outputs more interpretable and comparable with other CVIs but in vain for most CVIs. There were few CVIs with which we could do it and make experiments that suggest that this extension is relevant, but this is ongoing work and outside the scope of this JOSS submission.

### Conclusion

Problem number 3 is the problem that makes this feature impossible to implement for all CVIs. One possibility would be to make this feature possible for only compatible CVIs but this would probably discourage users to use the other CVIs, which is also something we would rather avoid. I am however willing to do it if you think that missing this feature would actually discourage users to use the PyCVI library at all. Please let me know what you think about it.

In the meantime, to facilitate the use of mutliple CVIs at once and to encourage citations, I have now implemented the [`pycvi.cvi.CVIAggregator`](https://pycvi.readthedocs.io/en/latest/pycvi.cvi.html#pycvi.cvi.CVIAggregator) class that allows users to combine several CVIs of their choice and to easily select the best clustering using a majority vote rule. The `CVIAggregator` syntax to compute scores and to select the best clustering is completely identical to any CVI syntax to do the same thing. To illustrate this and to showcase this added feature, I have added an example in the documentation "[CVIAggregator: Combining CVIs](https://pycvi.readthedocs.io/en/latest/examples/cvi_aggregator.html)" and I have updated the paper.

## On how to make PyCVI more likely to be cited

- Stronger emphasis on time series in the paper and in the documentation.
- Made it possible to compute CVI values and select the best clustering based on an aggregation of CVIs and the majority vote rule. Added the `pycvi.cvi.CVIAggregator` class and an example featuring `pycvi.cvi.CVIAggregator`.
- Stronger emphasis in the paper on the fact that there is no public python implementation of these CVIs.
