### 2024-08-20 #6841 @lrnv
@nglm Word count for the paper doubled, why is that ? Is joss, we aim at 1000 words max and 1-to-2 pages papers, this is now way too much figures and text
