### 2024-08-20 #6841 @nglm
@lrnv This was a slow accumulation of a few additions that were suggested during the review process, notably:

- Explain more Variation of Information
- Emphasize more the time-series constraints and PyCVI's contributions on this aspect (compatibility with DTW, DBA and times series machine learning libraries like aeon or sktime) to better explain the need of PyCVI as opposed to just contributing to sklearn.
- Add one example of successful clustering and clustering selection (the previous examples were meant to motivate the need of variation of information, of further research on CVIs and clustering methods, of a careful choice of CVI when clustering a given dataset, etc) as well as showcasing the compatibility of PyCVI with time-series data and non-time series data.
- Add an additional feature to make it easier to combine multiple CVIs to select the best clustering in order to facilitate practical usage but also to make PyCVI more likely to be cited as opposed to citing the authors of the individual and original CVIs.

I will try to make it more concise, but to go back to < 1000 words, I guess some of the explanations will have to be skipped.
