+++ examples/select_k/output-select_k.txt
@@ -1,148 +1,148 @@
 
  ***** Non time-series data with KMeans ***** 
 
-Clusterings generated in: 0.34s
+Clusterings generated in: 0.23s
  ====== Hartigan_monotonous ====== 
-0 231.36984326672615
-1 119.41081589902664
-2 117.57797077788588
-3 74.79373418541567
-4 96.47024458686965
-5 57.84956074543553
-6 51.82406719497778
-7 60.40054054521822
-8 50.97707086576322
+0 231.31309486007345
+1 119.53017448266189
+2 117.61566423046895
+3 75.96832482873897
+4 95.0656001459015
+5 57.30656920145313
+6 51.687707164140214
+7 59.10992898474898
+8 62.139148832813106
 9 None
 Code executed in 0.01 s
 Selected k: 1 | True k: 3
  ====== CalinskiHarabasz_original ====== 
 0 None
 1 0.0
-2 119.41081589902664
-3 141.9298657398708
-4 143.14019753113297
-5 166.21583557094027
-6 170.25495405112775
-7 175.12696558469432
-8 189.27359102849127
-9 200.42955766748833
-Code executed in 0.01 s
+2 119.53017448266186
+3 142.03940200451063
+4 143.99871868455529
+5 166.202708270155
+6 169.8882540473872
+7 174.67976151161565
+8 187.9661510034123
+9 206.79303061701037
+Code executed in 0.00 s
 Selected k: 9 | True k: 3
  ====== GapStatistic_original ====== 
 0 None
-1 -0.018247859414485923
-2 -0.16812365300209375
-3 -0.31493189672569155
-4 -0.5402252213606245
-5 -0.4316261196923472
-6 -0.42706979646402177
-7 -0.4524723582450232
-8 -0.41717508274232973
-9 -0.40213860370085497
-Code executed in 1.32 s
+1 0.006792153575967852
+2 -0.15580672581154076
+3 -0.3206999727877289
+4 -0.514634591757889
+5 -0.43655796457858465
+6 -0.43133194253892704
+7 -0.4440182169815916
+8 -0.4464424904973012
+9 -0.36545923888575427
+Code executed in 1.48 s
 Selected k: 1 | True k: 3
  ====== Silhouette ====== 
 0 None
 1 None
-2 0.29408441780040007
-3 0.31259123245038284
-4 0.31520954596824785
-5 0.37528288508948837
-6 0.37821693732242756
-7 0.41230912666282776
-8 0.42920512154344015
-9 0.4446931210382762
-Code executed in 0.31 s
+2 0.2942832123441149
+3 0.31865535671016215
+4 0.30974666695744546
+5 0.3758686461703666
+6 0.40405671642495117
+7 0.4088473534363824
+8 0.42222482515851223
+9 0.4565121748825993
+Code executed in 0.40 s
 Selected k: 9 | True k: 3
  ====== ScoreFunction ====== 
 0 None
 1 0.2674380756639073
-2 0.15435459454470002
-3 0.08425097006376447
-4 0.04721595655610222
-5 0.026516406474104692
-6 0.01938698432627184
-7 0.015157999822004053
-8 0.010427905172248786
-9 0.009228524453630427
-Code executed in 0.00 s
+2 0.1551856671096168
+3 0.08708406007354763
+4 0.04494313141469164
+5 0.026431295225813223
+6 0.019498243943131932
+7 0.01490000324244145
+8 0.01017110630012763
+9 0.010637054270650426
+Code executed in 0.01 s
 Selected k: 1 | True k: 3
  ====== MaulikBandyopadhyay_absolute ====== 
[…]
