### 2024-08-22 #6841 @nglm
@lrnv I have updated the figures.
(And sorry for spamming with the bot)
