+++ examples/utils.py
@@ -1,536 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-from matplotlib.cm import get_cmap
-from typing import List, Tuple, Dict, Any
-from math import ceil
-
-from pycvi.cvi import CVIs
-from pycvi.cluster import get_clustering
-
-
-def _get_nrows_ncols(nplots: int = None):
-    """
-    Adapt the figures to the total number of CVIs.
-
-    We want to know before creating the figure how many rows and columns
-    we will need. This depends on how many different k will be selected
-    and potentially if we have additional plots for the true data.
-
-    Parameters
-    ----------
-    nplots : int, optional
-        Number of plots, by default None, resulting in `nplots=len(CVIs)
-        + 2`
-    """
-    if nplots is None:
-        nplots = len(CVIs) + 2
-    n_rows = ceil(nplots / 5)
-    n_cols = 5
-    figsize = (4*n_cols, ceil(4*n_rows))
-    return n_rows, n_cols, figsize
-
-def _get_shape_UCR(data: np.ndarray) -> Tuple[Tuple[int], bool]:
-    """
-    Get the shape (N, T, d) of data and whether it is time series data.
-
-    Parameters
-    ----------
-    data : np.ndarray
-        The original data
-
-    Returns
-    -------
-    Tuple[Tuple[int], bool]
-        the shape (N, T, d) and whether it is time series data
-    """
-    dims = data.shape
-    if len(dims) == 3:
-        (N, T, d) = data.shape
-        UCR = True
-    else:
-        (N, d) = data.shape
-        T = 1
-        UCR = False
-    return (N, T, d), UCR
-
-def _get_colors(name: str="Set1") -> List:
-    """
-    Helper function to get a list of colors
-
-    Parameters
-    ----------
-    name : str, optional
-        Name of the matplotlib cmap, by default "Set1".
-
-    Returns
-    -------
-    List
-        A list of colors
-    """
-    cmap = get_cmap(name)
-    colors = cmap.colors
-    return colors
-
-def plot_cluster(
-    ax,
-    data: np.ndarray,
-    cluster: List[int],
-    color,
-):
-    """
-    Plot a given cluster on the given ax.
-
-    Works with UCR data (plot lines), and with non-time series data with
-    dimensions d = 1,2 or 3.
-
-    In case it is UCR data, use "color" for each line representing a
-    datapoint in the cluster.
-
-    Parameters
-    ----------
-    ax : A matplotlib axes
-        Where to plot the cluster
-    data : np.ndarray
-        The dataset
-    cluster : List[int]
-        The indices representing the cluster
-    color : _type_
-        The color to use to plot the cluster
-
-    Returns
-    -------
-    A matplotlib axes
-        The same matplotlib axes, but with the cluster plotted.
-    """
-    # Get the full shape and whether it is time-series data.
-    (N, T, d), UCR = _get_shape_UCR(data)
-
-    # If UCR, use plot type of plots.
-    if UCR:
-        # Transparency
-        alpha = 0.2
-        x = np.arange(T)
-        y_val = data[cluster, :, 0]
-
-        # Plot lines one by one, with the same color.
-        for y in y_val:
-            ax.plot(x, y, c=color, alpha=alpha)
-
-    # If non time series data, use scatter plots.
-    else:
-        # Size of the dots
-        s = 2
-        if d == 1:
-            x_val = np.zeros_like(data[cluster, 0])
-            y_val = data[cluster, 0]
-            ax.scatter(x_val, y_val, s=s)
-        elif d == 2:
-            x_val = data[cluster, 0]
-            y_val = data[cluster, 1]
-            ax.scatter(x_val, y_val, s=s)
-        elif d == 3:
-            x_val = data[cluster, 0]
[…]
