+++ docs/examples/select_k.rst
@@ -20,9 +20,9 @@ Here we assume that we are in real conditions, which means that we don't have ac
 
 .. image:: ../../examples/select_k/Barton_data_AgglomerativeClustering_Single.png
 
-.. image:: ../../examples/select_k/UCR_data_no_DTW_KMedoids-histogram.png
+.. image:: ../../examples/select_k/UCR_data_no_DTW_KMeans-histogram.png
 
-.. image:: ../../examples/select_k/UCR_data_no_DTW_KMedoids.png
+.. image:: ../../examples/select_k/UCR_data_no_DTW_KMeans.png
 
 .. image:: ../../examples/select_k/UCR_data_DTW_TimeSeriesKMeans-histogram.png
 
