### 2024-09-24 efea331
back to simpler pyproject
