+++ requirements.txt
@@ -1,19 +1,22 @@
-aeon==0.6.0 ; python_version >= "3.8" and python_version < "3.12"
-attrs==23.1.0 ; python_version >= "3.8" and python_version < "3.12"
+aeon==0.11.1 ; python_version >= "3.9" and python_version < "3.12"
+aeon==0.9.0 ; python_version >= "3.8" and python_version < "3.9"
 deprecated==1.2.14 ; python_version >= "3.8" and python_version < "3.12"
-importlib-metadata==7.0.0 ; python_version >= "3.8" and python_version < "3.9"
-joblib==1.3.2 ; python_version >= "3.8" and python_version < "3.12"
-llvmlite==0.41.1 ; python_version >= "3.8" and python_version < "3.12"
-numba==0.58.1 ; python_version >= "3.8" and python_version < "3.12"
+importlib-metadata==8.5.0 ; python_version >= "3.8" and python_version < "3.9"
+joblib==1.4.2 ; python_version >= "3.8" and python_version < "3.12"
+llvmlite==0.41.1 ; python_version >= "3.8" and python_version < "3.9"
+llvmlite==0.43.0 ; python_version >= "3.9" and python_version < "3.12"
+numba==0.58.1 ; python_version >= "3.8" and python_version < "3.9"
+numba==0.60.0 ; python_version >= "3.9" and python_version < "3.12"
 numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.12"
-packaging==23.2 ; python_version >= "3.8" and python_version < "3.12"
+packaging==24.1 ; python_version >= "3.8" and python_version < "3.12"
 pandas==2.0.3 ; python_version >= "3.8" and python_version < "3.12"
-python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.12"
-pytz==2023.3.post1 ; python_version >= "3.8" and python_version < "3.12"
+python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "3.12"
+pytz==2024.2 ; python_version >= "3.8" and python_version < "3.12"
 scikit-learn==1.3.2 ; python_version >= "3.8" and python_version < "3.12"
 scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.12"
 six==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
-threadpoolctl==3.2.0 ; python_version >= "3.8" and python_version < "3.12"
-tzdata==2023.3 ; python_version >= "3.8" and python_version < "3.12"
+threadpoolctl==3.5.0 ; python_version >= "3.8" and python_version < "3.12"
+typing-extensions==4.12.2 ; python_version >= "3.8" and python_version < "3.12"
+tzdata==2024.2 ; python_version >= "3.8" and python_version < "3.12"
 wrapt==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
-zipp==3.17.0 ; python_version >= "3.8" and python_version < "3.9"
+zipp==3.20.2 ; python_version >= "3.8" and python_version < "3.9"
