+++ docs/releases/PyCVI-0.1.4.md
@@ -1,0 +1,18 @@
+# PyCVI 0.1.4 release notes
+
+2024/09/24
+
+## Python versions
+
+This version supports Python versions 3.8 to 3.11.
+
+## Fixes
+
+- Some package compatibility issues fixed
+- Minor fixes in the documentation
+- JOSS paper fixed
+- GitHub actions updated
+
+## Contributors
+
+- Natacha Galmiche (@nglm)
