### 2024-09-25 #6841 @lrnv
@nglm Are you sure that there is no better citations (with DOI) for thigs such as tensorflow or skickitlearn ?
