+++ paper/paper.bib
@@ -73,6 +73,7 @@
 % time-series clustering and classification benchmark datasets
 @misc{UCRArchive2018,
   title = {The UCR Time Series Classification Archive},
+  doi = {10.48550/arXiv.1810.07758},
   author = {Dau, Hoang Anh and Keogh, Eamonn and Kamgar, Kaveh and Yeh, Chin-Chia Michael and Zhu, Yan
       and Gharghabi, Shaghayegh and Ratanamahatana, Chotirat Ann and Yanping and Hu, Bing
       and Begum, Nurjahan and Bagnall, Anthony and Mueen, Abdullah and Batista, Gustavo, and Hexagon-ML},
@@ -242,6 +243,7 @@
 % sklearn library paper
 @article{scikitlearn,
   title={Scikit-learn: Machine Learning in {P}ython},
+  doi={10.48550/arXiv.1201.0490},
   author={
     Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
     and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
@@ -352,10 +354,11 @@
 
 % Tensorflow official citation
 @misc{tensorflow2015,
-title={{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
-url={https://www.tensorflow.org/},
-note={Software available from tensorflow.org},
-author={
+  title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+  doi={10.5281/zenodo.4724125},
+  url={https://www.tensorflow.org/},
+  note={Software available from tensorflow.org},
+  author={
     Mart\'{i}n~Abadi and Ashish~Agarwal and Paul~Barham and Eugene~Brevdo
     and Zhifeng~Chen and Craig~Citro and Greg~S.~Corrado and Andy~Davis
     and Jeffrey~Dean and Matthieu~Devin and Sanjay~Ghemawat and Ian~Goodfellow
