### 2024-09-25 #6841 @nglm
@lrnv I have checked all the missing DOIs and could find DOIs for:
- The UCR Time Series Classification Archive
- Scikit-learn: Machine Learning in Python
- TensorFlow: Large-Scale Machine Learning on Hetero...
