+++ paper/paper.bib
@@ -202,11 +202,15 @@
 @article{lloyd1982least,
   title={Least squares quantization in PCM},
   author={Lloyd, Stuart},
+  url = {http://dx.doi.org/10.1109/tit.1982.1056489},
+  DOI = {10.1109/tit.1982.1056489},
   journal={IEEE transactions on information theory},
   volume={28},
+  ISSN = {0018-9448},
   number={2},
   pages={129--137},
   year={1982},
+  month = mar,
   publisher={IEEE}
 }
 
