+++ paper/paper.bib
@@ -26,7 +26,7 @@
   title={Clustering benchmarks},
   author={Barton, Tomas},
   year={2015},
-  howpublished = {"\url{https://github.com/deric/clusteringbenchmark}"},
+  howpublished = {"\url{https://github.com/deric/clustering-benchmark}"},
   note = "[Online; accessed 06-December-2023]",
   publisher={Github},
 }
