### 2024-09-26 #6841 @nglm
@lrnv Ok:
- Fixed link to  `clustering-benchmark` (https://github.com/deric/clustering-benchmark)
- referred to `scikit-learn-extra`, `aeon` and `clustering-benchmark` as links to repo
