### 2024-10-08 #6841 @crvernon
:wave: @nglm - I just need you to address the following before I accept this for publication.

**In your paper:**

- [ ] LINE 19:  The "p" in "python" should be capitalized.
- [ ] LINE 24:  The "p" in "python" should be capitalized.
- [ ] LINE 26:  Need as "s"  at the end of  "implementation" to make "implementations."
- [ ] LINES 27-28:  References should be in single parenthesis separated by a semicolon like "...(Arbelaitz et al., 2013; Gagolewski et al., 2021; Gurrutxaga et al., 2011; Theodoridis & Koutroumbas, 2009)..."
- [ ] Should LINE 33-43 be a single paragraph?
- [ ] LINE 52:  The "p" in "python" should be capitalized.
- [ ] LINE 55:  Should be "We experimented with 3 cases:"  where "with" is added in.
- [ ] LINE 65:  "meaning computed clusterings of poor quality." should be "meaning computed clusterings are of poor quality."  where "are" is added in.
- [ ] LINE 93:  Missing period at the end of the sentence.
- [ ] LINE 112:  The "r", as in the programming language, needs to be capitalized. You can do this in the bib file by placing characters that you wish to maintain the formatting of in curly brackets.

There is no need to create an additional release after making these changes.  Let me know when you are done and I'll take another look.  Thanks!
