+++ paper/paper.bib
@@ -65,7 +65,7 @@
   volume = {61},
   number = {6},
   author = {Malika Charrad and Nadia Ghazzali and V{\'{e}}ronique Boiteau and Azam Niknafs},
-  title = {NbClust: an R package for determining the relevant number of clusters in a data set},
+  title = {NbClust: an {R} package for determining the relevant number of clusters in a data set},
   journal = {Journal of Statistical Software}
 }
 
