### 2024-10-08 #6841 @crvernon
🥳 Congratulations on your new publication @nglm! Many thanks to @lrnv for editing and @robcaulk, @wob86, and @laurent-vouriot for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
